transactions: move the satisfaction test in from the lib root
